Reorder functions related to article update
