+++ src/clients/artikkelit/artikkelit.js
@@ -1,29 +1,17 @@
 import {showSnackbar, showTab} from '/common/ui-utils.js';
 import {Account, doLogin, logout} from '/common/auth.js';
-import {generateArticleRecord} from '/common/rest.js';
-import {showRecord} from '/common/marc-record-ui.js';
-import {
-  idbClear, idbDel, idbGet, idbSet,
-  idbGetStoredValues, getTableNames
-} from '/artikkelit/indexDB.js';
+import {idbClear, getTableNames} from '/artikkelit/indexDB.js';
 import {} from './articleActions.js';
-
-import {initAbstracts, refreshAbstractList} from '/artikkelit/interfaces/abstracts.js';
-import {
-  initAdditionalFields, refreshNotesList, refreshOtherTitlesList,
-  refreshUDKsList, refreshOtherRatingsList
-} from '/artikkelit/interfaces/additionalFields.js';
-
+import {initAbstracts} from '/artikkelit/interfaces/abstracts.js';
+import {initAdditionalFields} from '/artikkelit/interfaces/additionalFields.js';
 import {initArticle, resetAndHideCcLicense} from '/artikkelit/interfaces/article.js';
-import {initSciencesAndMethodologies, refreshSciencesList, refreshMetodologysList} from '/artikkelit/interfaces/sciencesAndMethodologies.js';
-
-import {initAuthors, refreshAuthorsList, refreshAuthorOrganizationList} from '/artikkelit/interfaces/authors.js';
-import {journalTemplate, bookTemplate} from '/artikkelit/constants/index.js';
+import {initSciencesAndMethodologies} from '/artikkelit/interfaces/sciencesAndMethodologies.js';
+import {initAuthors} from '/artikkelit/interfaces/authors.js';
 import {fillFormOptions, fillDatalistOptions, fillArticleTypeOptions} from '/artikkelit/loadData.js';
-import {initOntologyWords, ontologyTypeChange, refreshOntologyWordList} from '/artikkelit/interfaces/ontologyWords.js';
+import {initOntologyWords, ontologyTypeChange} from '/artikkelit/interfaces/ontologyWords.js';
 import {initPublicationSearch, resetPublicationSearchResultSelect} from '/artikkelit/interfaces/publicationSearch.js';
-import {initReviewSearch, resetReviewSearchResultSelect, refreshReviewsList} from '/artikkelit/interfaces/reviewSearch.js';
-import {resetCheckAndSave} from './articleActions.js';
+import {initReviewSearch, resetReviewSearchResultSelect} from '/artikkelit/interfaces/reviewSearch.js';
+import {refreshAllLists} from '/artikkelit/articleUpdate.js';
 
 
 /*****************************************************************************/
@@ -125,128 +113,6 @@ function sourceTypeChange(event) {
   doUpdate();
 }
 
-/*****************************************************************************/
-//UPDATE
-
-function updateRecordPreview(record) {
-  showRecord(record, 'previewRecord', {}, 'artikkelit', false);
-  showRecord(record, 'dialogRecord', {}, 'artikkelit', false);
-  resetCheckAndSave();
-}
-
-window.doUpdate = (event) => {
-  event?.preventDefault();
-  const tietueIndex = document.getElementById('julkaisu-haku-tulos-lista').value;
-
-  idbClear('artoRecord');
-  collectReviewsCheck();
-
-  const promises = [
-    idbGetStoredValues('artoSciences'),
-    idbGetStoredValues('artoMetodologys'),
-    idbGetStoredValues('artoAuthors'),
-    idbGetStoredValues('artoOntologyWords'),
-    idbGetStoredValues('artoAbstracts'),
-    idbGetStoredValues('artoNotes'),
-    idbGetStoredValues('artoUDKs'),
-    idbGetStoredValues('artoOtherRatings'),
-    idbGetStoredValues('artoReviews')
-  ];
-
-  if (tietueIndex === '') {
-    const sourceType = document.getElementById('kuvailtava-kohde').value;
-    if (sourceType == 'journal') {
-      promises.unshift(journalTemplate);
-    }
-    if (sourceType == 'book') {
-      promises.unshift(bookTemplate);
-    }
-  } else {
-    promises.unshift(idbGet('artoSources', parseInt(tietueIndex)));
-  }
-
-  Promise.all(promises).then(([
-    source,
-    sciences,
-    metodologys,
-    authors,
-    ontologyWords,
-    abstracts,
-    notes,
-    udks,
-    otherRatings,
-    reviews
-  ]) => {
-    const formData = collectFormData();
-    generateArticleRecord({
-      source,
-      ...formData,
-      sciences,
-      metodologys,
-      authors,
-      ontologyWords,
-      abstracts,
-      notes,
-      udks,
-      otherRatings,
-      reviews
-    })
-      .then(({record}) => {
-        setRecordToIndexDB(record);
-        updateRecordPreview(record);
-      })
-      .catch((error) => {
-        console.log('Error while generating article record: ', error);
-      });
-  });
-};
-
-
-/*****************************************************************************/
-//COLLECT
-
-function collectReviewsCheck() {
-  const articleType = document.getElementById('artikkelin-tyyppi').value;
-  const excludeReviews = ['A1', 'A2', 'A3'].some(str => articleType.includes(str));
-
-  if (excludeReviews) {
-    idbClear('artoReviews');
-  }
-}
-
-export function collectFormData() {
-  const [iso6391, iso6392b, ui] = document.getElementById('artikkelin-kieli').value.split(';');
-  const links = [];
-  document.getElementsByName('artikkelin-linkki').forEach(el => links.push(el.value));
-
-  return {
-    journalNumber: {
-      publishingYear: document.getElementById(`numeron-vuosi`).value,
-      volume: document.getElementById(`numeron-vol`).value,
-      number: document.getElementById(`numeron-numero`).value,
-      pages: document.getElementById(`numeron-sivut`).value
-    },
-    article: {
-      title: document.getElementById(`artikkelin-otsikko`).value,
-      titleOther: document.getElementById(`artikkelin-muu-nimeke`).value,
-      language: {iso6391, iso6392b, ui},
-      link: links,
-      type: document.getElementById(`artikkelin-tyyppi`).value,
-      sectionOrColumn: document.getElementById(`artikkelin-osasto-toistuva`).value,
-      ccLicense: document.getElementById(`artikkelin-cc-lisenssi`).value
-    },
-    collecting: {
-      f589a: document.getElementById(`poimintatiedot-poimintakoodi598a`).value,
-      f599a: document.getElementById(`poimintatiedot-poimintakoodi599a`).value,
-      f599x: document.getElementById(`poimintatiedot-poimintakoodi599x`).value
-    }
-  };
-}
-
-function setRecordToIndexDB(record) {
-  idbSet('artoRecord', 'record', record);
-}
-
 
 /*****************************************************************************/
 //RESET
@@ -257,19 +123,6 @@ function idbClearAllTables() {
   }
 }
 
-function refreshAllLists() {
-  refreshAbstractList();
-  refreshAuthorOrganizationList();
-  refreshAuthorsList();
-  refreshMetodologysList();
-  refreshNotesList();
-  refreshOtherTitlesList();
-  refreshOntologyWordList();
-  refreshOtherRatingsList();
-  refreshSciencesList();
-  refreshUDKsList();
-  refreshReviewsList();
-}
 
 function resetInputFields() {
   for (const inputField of document.getElementsByTagName('input')) {
